Update the type of TransactionNumber
TransactionNumber was previously defined as a generic int type. This is
not consistent with the rest of the codebase, which uses the int64 type.
This commit updates the type of TransactionNumber to int64, avoiding
casting.

+++ timo-trigger-driver-src/messages.go
@@ -4,7 +4,7 @@ import "time"
 
 type CommandMessage struct {
 	Command           string  `json:"command" bson:"command"`
-	TransactionNumber int     `json:"transaction_number,omitempty" bson:"transaction_number,omitempty"`
+	TransactionNumber int64   `json:"transaction_number,omitempty" bson:"transaction_number,omitempty"`
 	Userid            string  `json:"userid,omitempty" bson:"userid,omitempty"`
 	StockSymbol       string  `json:"stock_symbol,omitempty" bson:"stock_symbol,omitempty"`
 	Amount            float64 `json:"amount,omitempty" bson:"amount,omitempty"`
